Move the logic and functiosn to their own objects
Instead of a function based approach, the main logic is now moved to
appropriate objects like Certificate, which represents a single
Certificate and is backed by one CERT_CONTEXT, or a CertificateStore.

Also added the slf4j-simple package to have at least a very simple log
output of the example program App.