Add possiblity to connect certificate filters with logical AND, OR, NOT
That also generalises the implementation of list based certificate
filters. They can now be connected differently inside one filter. E.g. one
filter may require, that all properties of a list must be part of the
certificate property (AND), or only one of them (OR), which is the
default, or taht none of the properties must be part of the certificate
(NOT).