Simplify code by using Interface-driven approach of JNA
The native methods of used libraries are now defined in an interface with
the name of the library and made available through a pre-created INSTANCE
of this interface as a subclass of Library. This reduces the code size by
removing the need to getting the Funktion manually and invoking it with an
object of parameters. It also reduces the need of manually casting results
to specific types, this is now handled by the JNA library.